repackage: 쪽지 기록의 패키지를 변경한다
