+++ backend/src/main/java/com/now/naaga/letter/domain/letterlog/ReadLetterLog.java
@@ -1,4 +1,4 @@
-package com.now.naaga.letterlog.domain;
+package com.now.naaga.letter.domain.letterlog;
 
 import com.now.naaga.common.domain.BaseEntity;
 import com.now.naaga.game.domain.Game;
@@ -8,7 +8,7 @@ import jakarta.persistence.*;
 import java.util.Objects;
 
 @Entity
-public class WriteLetterLog extends BaseEntity {
+public class ReadLetterLog extends BaseEntity {
 
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
@@ -18,21 +18,21 @@ public class WriteLetterLog extends BaseEntity {
     @JoinColumn(name = "game_id")
     private Game game;
 
-    @OneToOne(fetch = FetchType.LAZY)
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "letter_id")
     private Letter letter;
 
-    protected WriteLetterLog() {
+    protected ReadLetterLog() {
     }
 
-    public WriteLetterLog(final Game game,
-                          final Letter letter) {
+    public ReadLetterLog(final Game game,
+                         final Letter letter) {
         this(null, game, letter);
     }
 
-    public WriteLetterLog(final Long id,
-                          final Game game,
-                          final Letter letter) {
+    public ReadLetterLog(final Long id,
+                         final Game game,
+                         final Letter letter) {
         this.id = id;
         this.game = game;
         this.letter = letter;
@@ -52,10 +52,14 @@ public class WriteLetterLog extends BaseEntity {
 
     @Override
     public boolean equals(final Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        WriteLetterLog that = (WriteLetterLog) o;
-        return Objects.equals(id, that.id);
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        ReadLetterLog letterLog = (ReadLetterLog) o;
+        return Objects.equals(id, letterLog.id);
     }
 
     @Override
@@ -65,7 +69,7 @@ public class WriteLetterLog extends BaseEntity {
 
     @Override
     public String toString() {
-        return "WriteLetterLog{" +
+        return "LetterLog{" +
                 "id=" + id +
                 ", gameId=" + game.getId() +
                 ", letterId=" + letter.getId() +
