+++ backend/src/main/java/com/now/naaga/letter/presentation/LetterLogController.java
@@ -1,4 +1,4 @@
-package com.now.naaga.letterlog.controller;
+package com.now.naaga.letter.presentation;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
