+++ backend/src/test/java/com/now/naaga/common/builder/ReadLetterLogBuilder.java
@@ -2,8 +2,8 @@ package com.now.naaga.common.builder;
 
 import com.now.naaga.game.domain.Game;
 import com.now.naaga.letter.domain.Letter;
-import com.now.naaga.letterlog.domain.ReadLetterLog;
-import com.now.naaga.letterlog.repository.ReadLetterLogRepository;
+import com.now.naaga.letter.domain.letterlog.ReadLetterLog;
+import com.now.naaga.letter.repository.letterlog.ReadLetterLogRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
