+++ backend/src/test/java/com/now/naaga/common/fixture/WriteLetterLogFixture.java
@@ -1,6 +1,6 @@
 package com.now.naaga.common.fixture;
 
-import com.now.naaga.letterlog.domain.WriteLetterLog;
+import com.now.naaga.letter.domain.letterlog.WriteLetterLog;
 
 import static com.now.naaga.common.fixture.GameFixture.GAME;
 import static com.now.naaga.common.fixture.LetterFixture.LETTER;
